+++ src/scheduler/Pricing.h
@@ -34,7 +34,7 @@ public:
   // Returns the second within the interval in which we should turn it on
   unsigned long getBestTime(Schedule& schedule);
 
-  uint32_t getPrice(unsigned long day, uint8_t hour);
+  uint8_t getPrice(unsigned long day, uint8_t hour);
 
 //private:
   bool getTomorrowPrices(unsigned long date);
@@ -45,7 +45,7 @@ public:
   // 0-3 prices of the day before
   // 4-27 prices of the next day
   // it is updated every day at ~20:00
-  uint32_t _price[28];
+  uint8_t _price[28];
   unsigned long _date;
 };
 #endif
